Rename components to pageobjects
